Move FR_KNOWN_TOKENS_CACHE to Spellcheck module
