+++ robotoff/utils/text.py
@@ -49,12 +49,3 @@ def get_nlp(lang: str):
 FR_NLP_CACHE = cache.CachedStore(
     functools.partial(get_nlp, lang="fr"), expiration_interval=None
 )
-
-
-def get_fr_known_tokens() -> Set[str]:
-    tokens = set(text_file_iter(settings.INGREDIENT_TOKENS_PATH, comment=False))
-    tokens = tokens.union(set(text_file_iter(settings.FR_TOKENS_PATH, comment=False)))
-    return tokens
-
-
-FR_KNOWN_TOKENS_CACHE = cache.CachedStore(get_fr_known_tokens)
